refactor : 폴더구조 통일
