+++ src/assets/images/detail/sample/cover.png
