+++ src/components/_common/Degree.js
